+++ src/types.ts
@@ -37,7 +37,6 @@ export type {
   CollectionEncryption,
   CollectionEncryptionEpoch,
   CollectionEncryptionEpochsMac,
-  CollectionEncryptionEpochsSig,
   CollectionEncryptionRecipient,
   PolicyDocument,
   LinkSet,